more code clean tests and overrides
