--- .yaml-lint.yml
@@ -1,10 +1,0 @@
----
-extends: default
-ignore: |
-  molecule_kubevirt/test/refs
-  # .tox
-  # .git
-  # vault.yml
-  # roles-dependencies
-  # collections-dependencies
-  # molecule
