+++ molecule_kubevirt/test/refs/instance-full.yml
@@ -50,8 +50,10 @@ spec:
           type: q35
         resources:
           limits:
-            memory: 2Gi
+            cpu: "1"
+            memory: 3Gi
           requests:
+            cpu: 200m
             memory: 2Gi
       hostname: myhost
       networks:
